Pass username/password to client
Currently when python clients are called, Rally authenticates
with keystone client then pass then token to the clients.
This behavior will break the test when the token gets expired.

When the token gets expired, the operations using the token
will receive a 401 error which instructs the clients to
re-new the token with the user crendentials. The python clients
will re-authenticate and try to re-send the operation request.

However, since the credentials are not passed by Rally,
the clients are unable to re-authenticate in which case
the operation will fail and break the test.

This patch is going to pass the credentials to the python clients
to allow the test go beyond the token expiration time.

This patch mainly focus on nova and neutron.
Please note that some clients do not support this
re-authentication mechanism.

Change-Id: I2732e9c7053005b265df3122181f99544d4b6100