Add sleep_between in live migration scenarios
Main problem when using live migration is running it on a utilized VM.
To simulate this case when benchmarking tester need to spawn a VM that
actually does something. The stressing application may need some time
to spin up. This commit adds configurable sleep_between to live
migration scenarios between booting and live migrating a VM to let
tests wait for eventual stress app to start.

Change-Id: I807d240d93fb42e91e9ce096e328039a6f5d1456