Add Nova scenario to boot multiple servers in one request
The performance profile of, e.g., creating 10 VMs in a single request
is often very, very different from creating 10 VMs in 10 requests, and
the difference only grows with the number of VMs. This adds a scenario
to boot an arbitrary number of servers with a single request, which
can be compared to boot_and_delete_server, which boots them one server
per request.

Change-Id: I8cb5d6a4f27cce228b83d6a869ed75c505dd6626