Remove _delete_all_servers() nova utils function
This function deletes all servers in the current tenant, which is
almost certainly The Wrong Thing. It's also poorly written: it deletes
the servers sequentially, not in parallel; and it creates multiple
timers for the same action. A subsequent commit will add support for
deleting multiple specific servers at once.

Change-Id: Ib381fcc93d77bf417ce7d90c90261ca0209c1bbb