Better mocking for self.[admin_]clients in Scenario tests
self.clients and self.admin_clients are mocked out in all sorts of fun
and interesting ways in the Scenario test cases, but rarely is it done
the same way twice. Many of the tests fail to distinguish between a
call to self.clients and a call to self.admin_clients, and all (or
nearly all) don't distinguish between clients -- that is, if a test
expects a call to self.clients("nova").foo, that would be satisfied by
self.clients("keystone").foo. This patch fixes those issues, and gives
us a consistent way to mock them going forward.

Change-Id: I98a532cedbe186c49bd5b6db903a72e7d1b6bcc7